feat: Add DeliveryRequest model and CRUD handlers; update TMS integration and fix type mismatches

+++ handler/tms_handler.go
@@ -54,7 +54,7 @@ func GetTranportRequestsHandler(ctx *gin.Context) {
 		logger.Error(error)
 		return
 	}
-	trs, error := tmsClient.GetNodeTransportRequests(nodeId)
+	trs, error := tmsClient.GetNodeTransportRequests(uint(nodeId))
 	trResp := make([]db.TransportRequest, len(trs))
 	for i := range trs {
 		trResp[i] = db.TransportRequest{
